packages/systemrdl-lsp/pyproject.toml
[project]
name = "systemrdl-lsp"
version = "0.21.0"
description = "Language Server Protocol implementation for SystemRDL 2.0 — diagnostics, hover, goto-def, completion, semantic tokens, and the full LSP feature surface (powered by systemrdl-compiler)"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "seimei-d" }]
keywords = ["systemrdl", "lsp", "language-server", "hardware", "registers"]
classifiers = [
  "Development Status :: 4 - Beta",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Topic :: Software Development :: Compilers",
  "Topic :: Text Editors :: Integrated Development Environments (IDE)",
]
dependencies = [
  "pygls>=1.3.1",
  "systemrdl-compiler>=1.29.0",
  "lsprotocol>=2024.0.0b1",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-asyncio>=0.23", "ruff>=0.6"]

[project.scripts]
systemrdl-lsp = "systemrdl_lsp.__main__:main"

[project.urls]
Homepage = "https://github.com/seimei-d/systemrdl-pro"
Issues = "https://github.com/seimei-d/systemrdl-pro/issues"
Source = "https://github.com/seimei-d/systemrdl-pro/tree/main/packages/systemrdl-lsp"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/systemrdl_lsp"]
